added experimental DigitalOutput builder for evaluation/review.

+++ plugins/pi4j-plugin-pigpio/src/main/java/com/pi4j/plugin/pigpio/PiGpioPlugin.java
@@ -110,6 +110,6 @@ public class PiGpioPlugin implements Plugin {
     @Override
     public void shutdown(Context context) throws IOException {
         // shutdown the PiGpio library
-        piGpio.terminate();
+        if(piGpio != null) piGpio.terminate();
     }
 }
